Fireballdmg goldscaling

+++ The Dark Woods master/Assets/Resources/Scripts/Player/FireBallLogic.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class FireBallLogic : MonoBehaviour
 {
     //Variables
-    public int Damage = 10;
+    public float Damage = 10;
     //Unity functions
     public void Update()
     {
